类型强转

+++ kernel/sysfile.c
@@ -560,7 +560,7 @@ uint64 sys_munmap(void){
         pr->areaps[i] = 0;
         return 0;
       }else {
-        uvmunmap(pr->pagetable, pr->areaps[i]->addr, length/PGSIZE, 1);
+        uvmunmap(pr->pagetable, (uint64)pr->areaps[i]->addr, length/PGSIZE, 1);
         pr->areaps[i]->addr += length;
         pr->areaps[i]->length -= length;
         return 0;
